cifutils/pyproject.toml
[project]
name = "cifutils"
version = "2.0.0"
description = "CIFUtils: Python packages for reading mmCIF files from the PDB"
readme = "README.md"
requires-python = ">= 3.8"
authors = [
    { name = "Nate Corley", email = "ncorley@uw.edu" },
    { name = "Simon Mathis", email = "smathis@uw.edu" },
    { name = "Pascal Strumfels"},
    { name = "Rohith Krishna"},
]
dependencies = [
    # NOTE: Dependency versions specify the minimum versions for the packages 
    #  that were confirmed to work with the codebase. Lower versions might still
    #  work (at your own risk). 
    "pandas>=1.4.2,<2",
    "numpy>=1.25.0,<2",
    "cytoolz>=0.12.3,<1",
    "biopython>=1.83,<2",
    "fire>=0.6.0,<1",
]

[project.optional-dependencies]
dev = [
    "ruff==0.6.2",
    "pre-commit==3.7.1",
    "pytest>=8.2.0,<9",
    "pytest-dotenv>=0.5.2,<1",
    "pytest-cov>=4.1.0,<5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["cifutils"]

[tool.hatch.build.targets.wheel.force-include]
# to include data files in the wheel when `pip install`-ing (if we want to give some test cases)

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "env",
    ".venv",
    "venv",
    "data"
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
docstring-code-format = true
docstring-code-line-length = 88

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Avoid linting notebooks
exclude = ["*.ipynb"]

[tool.pytest]
conftest = "tests/conftest.py"